requirements.txt
# Core ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
xgboost>=1.7.0
joblib>=1.3.0

# Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
